requirements.txt
streamlit==1.26.0
SQLAlchemy==2.0.29
psycopg[binary]==3.2.1
python-dotenv==1.0.0
pytest==7.4.0
